requirements-crypto.txt
# 虛擬貨幣期貨交易系統依賴套件

# 核心依賴
websocket-client>=1.7.0  # WebSocket 連接
requests>=2.31.0         # HTTP 請求

# 數據處理（可選）
numpy>=1.24.0           # 數值計算
pandas>=2.0.0           # 數據分析

# 技術指標（可選 - 用於進階策略）
ta-lib>=0.4.0           # 技術分析庫
# 注意：ta-lib 需要額外安裝 C 庫
# Windows: 從 https://www.lfd.uci.edu/~gohlke/pythonlibs/#ta-lib 下載 whl 文件
# Linux/Mac: 先安裝 ta-lib C 庫，然後 pip install ta-lib

# 機器學習（可選 - 用於高級 AI 策略）
# scikit-learn>=1.3.0
# tensorflow>=2.13.0
# torch>=2.0.0

# 可視化（可選）
# matplotlib>=3.7.0
# plotly>=5.14.0

# 通知功能（可選）
# python-telegram-bot>=20.0  # Telegram 通知
# sendgrid>=6.10.0           # Email 通知

# 日誌和監控（可選）
# loguru>=0.7.0              # 更好的日誌系統
# prometheus-client>=0.16.0   # 監控指標
